enigma2 20120828 master -> 20120912 master ------------------------------------------ - fixed signal loss after closing linked tuner - fixed removal of temporary opkg files - fixed servicelist / transponderlist ordering (fixes autotimer) - added support for transparent GIF files - fixed audio/video loss after service change (after starting background applications)

+++ usr/include/enigma2/lib/dvb/idvb.h
@@ -145,8 +145,8 @@ public:
 	eDVBNamespace(): v(-1) { }
 	bool operator == (const eDVBNamespace &c) const { return v == c.v; }
 	bool operator != (const eDVBNamespace &c) const { return v != c.v; }
-	bool operator < (const eDVBNamespace &c) const { return v < c.v; }
-	bool operator > (const eDVBNamespace &c) const { return v > c.v; }
+	bool operator < (const eDVBNamespace &c) const { return (unsigned int)v < (unsigned int)c.v; }
+	bool operator > (const eDVBNamespace &c) const { return (unsigned int)v > (unsigned int)c.v; }
 };
 
 struct eDVBChannelID
